requirements.txt
requests
python-telegram-bot
openai
python-dotenv
httpx
pysocks
